✨ Feat: ProfileCover구조 변경
