--- src/pages/profile/components/ProfileBackground/ProfileCover.tsx
@@ -1,12 +1,0 @@
-import { ProfileCoverImage } from './ProfileCover.style';
-
-interface ProfileBackgroundProps {
-  src: string;
-  children?: React.ReactNode;
-}
-
-const ProfileCover = ({ src, children }: ProfileBackgroundProps) => {
-  return <ProfileCoverImage src={src}>{children}</ProfileCoverImage>;
-};
-
-export default ProfileCover;
