+++ src/pages/profile/components/ProfileCover/index.ts
